--- a/.teamcity/_self/projects/WebApp.kt
+++ b/.teamcity/_self/projects/WebApp.kt
@@ -598,8 +598,8 @@ fun playwrightPrBuildType( targetDevice: String, buildUuid: String ): BuildType 
 					export NODE_ENV="test"
 					export PLAYWRIGHT_BROWSERS_PATH=0
 
-					# Install modules
-					${_self.yarn_install_cmd}
+					# Install deps
+					yarn workspaces focus wp-e2e-tests wp-e2e-tests
 
 					# Build packages
 					yarn workspace @automattic/calypso-e2e build