--- a/test/e2e/package.json
+++ b/test/e2e/package.json
@@ -22,6 +22,7 @@
 		"@automattic/calypso-typescript-config": "^1.0.0"
 	},
 	"dependencies": {
+		"@automattic/calypso-build": "^9.0.0",
 		"@automattic/calypso-e2e": "^0.1.0",
 		"@automattic/mocha-debug-reporter": "^0.1.0",
 		"@automattic/testarmada-magellan-mocha-plugin": "^10.0.0",