#### Background

Playwright e2e builds in TeamCity spend 58 seconds on average in the step "Preparing environment". A huge portion of that time (47s) is just running `yarn install`.

#### Changes proposed in this Pull Request

Enables [Yarn focused mode](https://yarnpkg.com/cli/workspaces/focus) to limit the number of dependencies installed. We only install the dependencies for the workspaces `wp-e2e-tests` (aka `./test/e2e`) and `@automattic/calypso-e2e`. 

This brings down the "Preparing environment" time to 29s (17s for `yarn`).

An unfortunate side effect is that yarn will output a ton of warnings like:

```
@emotion/native ➤ peerDependencies ➤ @emotion/core: No matching package in the dependency tree; you may not need this rule anymore.
```

This is caused by the overrides in `.yarnrc.yaml`. They do apply to other workspaces, but as they are not the workspaces installed, Yarn think they don't.


#### Testing instructions

Check e2e tests still work